Move cmd to connect
